requirements.txt
PyQt6>=6.0.0
pyotp>=2.8.0
cryptography>=41.0.0
pillow>=10.0.0
qrcode>=7.4.2
pyperclip>=1.8.2
pyinstaller>=6.3.0